chore(aero-usb): fix clippy -D warnings
- Clean up test helpers to satisfy clippy::identity_op/erasing_op, bool_assert_comparison, field_reassign_with_default, default_constructed_unit_structs, and manual_div_ceil
- Simplify a few library helpers (contains/collapsible_if) and remove a useless vec allocation

Verified: cargo clippy -p aero-usb --all-targets --all-features -- -D warnings; cargo test -p aero-usb

+++ crates/aero-usb/tests/uhci_randomized_snapshot_roundtrip.rs
@@ -237,8 +237,8 @@ fn gen_write_op(rng: &mut Rng) -> Op {
                 }
                 v as u32
             }
-            regs::REG_FRNUM | REG_FRNUM_HI => (rng.next_u32() & 0x07ff) as u32,
-            regs::REG_SOFMOD => (rng.next_u32() & 0xff) as u32,
+            regs::REG_FRNUM | REG_FRNUM_HI => rng.next_u32() & 0x07ff,
+            regs::REG_SOFMOD => rng.next_u32() & 0xff,
             regs::REG_PORTSC1 | REG_PORTSC1_HI | regs::REG_PORTSC2 | REG_PORTSC2_HI => {
                 const CSC: u16 = 1 << 1;
                 const PED: u16 = 1 << 2;
